requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=15.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0.post1
fastapi>=0.110.0
uvicorn>=0.25.0
typer>=0.12.0
neo4j>=5.20.0
python-dotenv>=1.0.1

